updated casting for windows compile

+++ src/shapiq/tree/interventional/cext/cext.cc
@@ -1005,11 +1005,11 @@ static PyObject *compute_interactions_flatten(PyObject *self, PyObject *args)
         {
             for (int i = 0; i < n_iterations; i++)
             {
-                int feature = features[i];
+                int feature = static_cast<int>(features[i]);
                 double leaf_val = leaf_predictions[i];
-                int e = e_sizes_data[i];
-                int r = r_sizes_data[i];
-                int s_cap_e = feature_in_e_data[i];
+                int e = static_cast<int>(e_sizes_data[i]);
+                int r = static_cast<int>(r_sizes_data[i]);
+                int s_cap_e = static_cast<int>(feature_in_e_data[i]);
                 int s_cap_r = 1 - s_cap_e;
                 int s = 1;
                 int sign = (s_cap_r % 2 == 0) ? 1 : -1;
@@ -1042,11 +1042,11 @@ static PyObject *compute_interactions_flatten(PyObject *self, PyObject *args)
         {
             for (int i = 0; i < n_iterations; i++)
             {
-                int feature = features[i];
+                int feature = static_cast<int>(features[i]);
                 double leaf_val = leaf_predictions[i];
-                int e = e_sizes_data[i];
-                int r = r_sizes_data[i];
-                int s_cap_e_i = feature_in_e_data[i];
+                int e = static_cast<int>(e_sizes_data[i]);
+                int r = static_cast<int>(r_sizes_data[i]);
+                int s_cap_e_i = static_cast<int>(feature_in_e_data[i]);
                 int s_cap_r_i = 1 - s_cap_e_i;
                 int s = 1;
                 int sign = (s_cap_r_i % 2 == 0) ? 1 : -1;
@@ -1082,8 +1082,8 @@ static PyObject *compute_interactions_flatten(PyObject *self, PyObject *args)
                         break;
                     }
 
-                    int feature_j = features[j];
-                    int s_cap_e_j = feature_in_e_data[j];
+                    int feature_j = static_cast<int>(features[j]);
+                    int s_cap_e_j = static_cast<int>(feature_in_e_data[j]);
                     int s_cap_r_j = 1 - s_cap_e_j;
 
                     // Combined counts for the pair (i, j)
